UPDATES HF_COLIN to current

+++ common/cmd.c
@@ -29,7 +29,6 @@
  * @file cmd.c
  * @brief
  */
-
 #include "cmd.h"
 
 // (iceman 2017) this method is not used anymore. uart_win32 /uart_posix is used instead
@@ -51,7 +50,7 @@ bool cmd_receive(UsbCommand* cmd) {
   return (rxlen);
 }
 
-bool cmd_send(uint32_t cmd, uint32_t arg0, uint32_t arg1, uint32_t arg2, void* data, size_t len) {
+bool cmd_send(uint64_t cmd, uint64_t arg0, uint64_t arg1, uint64_t arg2, void* data, size_t len) {
 
 	UsbCommand txcmd;
 
@@ -79,5 +78,3 @@ bool cmd_send(uint32_t cmd, uint32_t arg0, uint32_t arg1, uint32_t arg2, void* d
 
 	return true;
 }
-
-
